requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
starlette
python-multipart
itsdangerous
jinja2